requirements.txt
cement==3.0.4
jinja2
pyyaml
colorlog
bs4
lxml
